+++ lib/terradactyl/terraform/commands/destroy.rb
@@ -80,6 +80,12 @@ module Terradactyl
       end
     end
 
+    module Rev1_05
+      module Destroy
+        include Terradactyl::Terraform::Rev015::Destroy
+      end
+    end
+
     module Commands
       class Destroy < Base
       end