+++ spec/fixtures/stacks/rev1_05/versions.tf
@@ -1,0 +1,9 @@
+terraform {
+  required_version = "~> 1.5.0"
+  required_providers {
+    null = {
+      version = "~> 3.0"
+      source  = "hashicorp/null"
+    }
+  }
+}